refactor(RegisterForm): add separation of concerns
